:white_check_mark: Basic unit testing.

+++ contracts/openticket/util/validation/Validations.sol
@@ -5,12 +5,7 @@ pragma solidity ^0.8.14;
 import "../../model/OpenTicketModel.sol";
 
 library Validations {
-
-    
-
-    function isValid(
-        OpenTicketModel.AdmissionEvent memory admissionEvent
-    )
+    function isValid(OpenTicketModel.AdmissionEvent memory admissionEvent)
         internal
         pure
         validEvent(admissionEvent)
@@ -19,38 +14,30 @@ library Validations {
     function isValidTicket(
         OpenTicketModel.AdmissionEvent memory admissionEvent,
         OpenTicketModel.OpenTicket memory ticket
-    )
-        internal
-        pure
-        validEvent(admissionEvent)
-        validTicket(ticket)
-    {}
+    ) internal pure validEvent(admissionEvent) validTicket(ticket) {}
 
-    function canBuy(
-        OpenTicketModel.OpenTicket memory ticket,
-        uint256 amount
-    )
+    function canBuy(OpenTicketModel.OpenTicket memory ticket, uint256 amount)
         internal
         validTicket(ticket)
-        hasEnoughMoney(ticket,amount)
-        hasEnoughSupply(ticket,amount)
+        hasEnoughMoney(ticket, amount)
+        hasEnoughSupply(ticket, amount)
     {}
 
-    function canWithdraw(
-        address account
-    )
-        internal
-        view
-        hasBalance(account)
-    {}
+    function canWithdraw(address account) internal view hasBalance(account) {}
 
     modifier validTicket(OpenTicketModel.OpenTicket memory ticket) {
         require(ticket.minted, "OT: ticket not minted");
         _;
     }
 
-    modifier hasEnoughMoney(OpenTicketModel.OpenTicket memory ticket, uint256 amount) {
-        require(msg.value >= (ticket.unitPrice * amount), "OT: not enough money");
+    modifier hasEnoughMoney(
+        OpenTicketModel.OpenTicket memory ticket,
+        uint256 amount
+    ) {
+        require(
+            msg.value >= (ticket.unitPrice * amount),
+            "OT: not enough money"
+        );
         _;
     }
 
@@ -59,7 +46,10 @@ library Validations {
         _;
     }
 
-    modifier hasEnoughSupply(OpenTicketModel.OpenTicket memory ticket, uint amount) {
+    modifier hasEnoughSupply(
+        OpenTicketModel.OpenTicket memory ticket,
+        uint256 amount
+    ) {
         require(amount <= ticket.supply, "OT: not enough supply");
         _;
     }
